requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
langchain-community
together
pydantic